+++ example/src/server/init.server.lua
@@ -1,0 +1,10 @@
+local ReplicatedStorage = game:GetService("ReplicatedStorage")
+local start = require(ReplicatedStorage.Shared.start)
+local setupTags = require(ReplicatedStorage.Shared.setupTags)
+
+local world = start({
+	script.systems,
+	ReplicatedStorage.Shared.systems,
+})
+
+setupTags(world)