+++ example/src/server/systems/replication.lua
@@ -1,6 +1,6 @@
 local Players = game:GetService("Players")
 local ReplicatedStorage = game:GetService("ReplicatedStorage")
-local Components = require(ReplicatedStorage.Game.components)
+local Components = require(ReplicatedStorage.Shared.components)
 local useEvent = require(ReplicatedStorage.Lib.Matter).useEvent
 
 local RemoteEvent = Instance.new("RemoteEvent")