+++ lib/Loop.lua
@@ -1,3 +1,4 @@
+local RunService = game:GetService("RunService")
 local topoRuntime = require(script.Parent.topoRuntime)
 local rollingAverage = require(script.Parent.rollingAverage)
 
@@ -62,6 +63,10 @@ function Loop.new(...)
 		_stateLength = select("#", ...),
 		_systemState = {},
 		_middlewares = {},
+		_systemErrors = {},
+		_systemLogs = {},
+		profiling = nil,
+		trackErrors = false,
 	}, Loop)
 end
 
@@ -129,6 +134,11 @@ function Loop:scheduleSystems(systems: { System })
 	for _, system in ipairs(systems) do
 		self._systems[system] = system
 		self._systemState[system] = {}
+
+		if RunService:IsStudio() then
+			-- In Studio, we start logging immediately.
+			self._systemLogs[system] = {}
+		end
 	end
 
 	self:_sortSystems()
@@ -157,12 +167,14 @@ function Loop:evictSystem(system: System)
 	end
 
 	self._systems[system] = nil
+	self._systemErrors[system] = nil
 
 	topoRuntime.start({
 		system = self._systemState[system],
 	}, function() end)
 
 	self._systemState[system] = nil
+	self._systemLogs[system] = nil
 
 	self:_sortSystems()
 end
@@ -216,7 +228,7 @@ local function orderSystemsByDependencies(unscheduledSystems: { System })
 
 	local scheduledSystemsSet = {}
 	local scheduledSystems = {}
-	local tombstone = {}
+	local tombstone: any = {}
 
 	while #scheduledSystems < #unscheduledSystems do
 		local atLeastOneScheduled = false
@@ -336,7 +348,7 @@ function Loop:begin(events)
 
 			generation = not generation
 
-			local dirtyWorlds = {}
+			local dirtyWorlds: {[any]: true} = {}
 			local profiling = self.profiling
 
 			for _, system in ipairs(self._orderedSystemsByEvent[eventName]) do
@@ -346,6 +358,7 @@ function Loop:begin(events)
 						generation = generation,
 						deltaTime = deltaTime,
 						dirtyWorlds = dirtyWorlds,
+						logs = self._systemLogs[system],
 					},
 					currentSystem = system,
 				}, function()
@@ -407,6 +420,28 @@ function Loop:begin(events)
 							task.spawn(error, errorString)
 							warn("Matter: The above error will be suppressed for the next 10 seconds")
 							recentErrors[errorString] = true
+						end
+
+						if self.trackErrors then
+							if self._systemErrors[system] == nil then
+								self._systemErrors[system] = {}
+							end
+
+							local errorStorage = self._systemErrors[system]
+							local lastError = errorStorage[#errorStorage]
+
+							if lastError and lastError.error == errorString then
+								lastError.when = os.time()
+							else
+								table.insert(errorStorage, {
+									error = errorString,
+									when = os.time(),
+								})
+
+								if #errorStorage > 100 then
+									table.remove(errorStorage, 1)
+								end
+							end
 						end
 					end