+++ lib/init.lua
@@ -70,6 +70,7 @@ return {
 	useEvent = require(script.hooks.useEvent),
 	useDeltaTime = require(script.hooks.useDeltaTime),
 	useThrottle = require(script.hooks.useThrottle),
+	log = require(script.hooks.log),
 	useHookState = topoRuntime.useHookState,
 	useCurrentSystem = topoRuntime.useCurrentSystem,